move models into lambda_handler
