+++ packages/shared-models/src/shared_models/__init__.py
@@ -2,7 +2,6 @@ from enum import StrEnum
 
 from pydantic import BaseModel
 from pydantic import ConfigDict
-from pydantic import Field
 
 
 class CdaInstanceIdentifier(BaseModel):
@@ -69,21 +68,3 @@ class TTCAugmenterInput(BaseModel):
     )
     eicr_id: str
     nonstandard_codes: list[NonstandardCodeInstance]
-
-
-class S3Location(BaseModel):
-    """Represents the location of a file in S3, indicating which file contained the relevant data."""
-
-    bucket: str = Field(description="The S3 bucket where the file is located.")
-    key: str = Field(description="The S3 key (path) where the file is located.")
-
-
-class OpenSearchHitSource(BaseModel):
-    """Represents a single search result _source returned from OpenSearch."""
-
-    id: int = Field(description="The unique ID from the embedding data of the search result hit.")
-    loinc_code: str = Field(description="The LOINC code of the search result hit.")
-    loinc_name_type: str = Field(description="The LOINC name type of the search result hit.")
-    description: str = Field(description="The description of the search result hit.")
-    loinc_type: str = Field(description="The LOINC type of the search result hit.")
-    s3: S3Location = Field(description="The S3 location of the search result hit.")
